Modification of the nemlumat.f file
A change was implemented in the nemlumat.f file. The multipliers for shear stress and strain, which convert them to Mandel notation, were incorrectly defined. The stress multipliers have been adjusted from 1.0/√2.0 to √2.0. Additionally, the shear strain components were reduced by half to align with Abaqus's interpretation of shear strains as engineering strains, not tensorial strains.